# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "EP1 Exam",
    "description": "documentation for Noroff EP - e-commerce"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets list of all categories",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Creates a new category",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/postcategory"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Alters existing category of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/altercategory"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Deletes category of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/delete"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/categories/{categoryid}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets list of category of id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "categoryid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Gets list of all brands",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "Gets list of all brands",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/postbrand"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "Gets brand of specified id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/alterbrand"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "Gets brand of specified id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/delete"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/brands/{brandid}": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Gets brand of specified id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "brandid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/membership": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Alter users membership manually",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/usermembership"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/memberships/": {
      "get": {
        "tags": [
          "Memberships"
        ],
        "description": "Gets all memberships",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Memberships"
        ],
        "description": "Creates new membership",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/postmembership"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Memberships"
        ],
        "description": "Creates new membership",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/altermembership"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Memberships"
        ],
        "description": "delete membership of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/delete"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/memberships/{membershipid}": {
      "get": {
        "tags": [
          "Memberships"
        ],
        "description": "Gets membership of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "membershipid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "postbrand": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string",
          "example": "Ontario sportswear"
        }
      },
      "required": [
        "brand"
      ]
    },
    "alterbrand": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "brand": {
          "type": "string",
          "example": "Ontario sportswear"
        }
      },
      "required": [
        "id",
        "brand"
      ]
    },
    "delete": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "id"
      ]
    },
...
    "postcategory": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "Hockey jerseys"
        }
      },
      "required": [
        "category"
      ]
    },
    "altercategory": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "Hockey jerseys"
        },
        "id": {
          "type": "number",
          "example": 69
        }
      },
      "required": [
        "category",
        "id"
      ]
    },
    "postmembership": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "NOSHO member"
        },
        "minItems": {
          "type": "number",
          "example": 69
        },
        "maxItems": {
          "type": "number",
          "example": 100
        },
        "discount": {
          "type": "number",
          "example": 60
        }
      },
      "required": [
        "name",
        "minItems",
        "maxItems",
        "discount"
      ]
    },
    "altermembership": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "NOSHO member"
        },
        "minItems": {
          "type": "number",
          "example": 69
        },
        "maxItems": {
          "type": "number",
          "example": 100
        },
        "discount": {
          "type": "number",
          "example": 60
        }
      },
      "required": [
        "id",
        "name",
        "minItems",
        "maxItems",
        "discount"
      ]
    },
...
    "usermembership": {
      "type": "object",
      "properties": {
        "userid": {
          "type": "number",
          "example": 2
        },
        "membershipid": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "userid",
        "membershipid"
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": {
    "Admin": [
      "adminJWTtoken"
    ]
  }
}